[UPDATE] Fixing imports after rename
